+++ tests/problems/conftest.py
@@ -33,53 +33,53 @@ if TYPE_CHECKING:
     from collections.abc import Mapping
 
     from gemseo.algos.design_space import DesignSpace
-    from gemseo_benchmark.problems.base_benchmarking_problem import (
-        BaseBenchmarkingProblem,
-    )
 
     from gemseo_benchmark.algorithms.algorithms_configurations import (
         AlgorithmsConfigurations,
     )
     from gemseo_benchmark.benchmarker.base_worker import BaseWorker
+    from gemseo_benchmark.problems.base_problem_configuration import (
+        BaseProblemConfiguration,
+    )
     from gemseo_benchmark.results.results import Results
 
 
 def check_no_starting_point(
-    benchmarking_problem: BaseBenchmarkingProblem,
+    problem_configuration: BaseProblemConfiguration,
 ) -> None:
     """Check that there is no starting point.
 
     Args:
-        benchmarking_problem: The benchmarking problem.
+        problem_configuration: The problem configuration.
     """
     with pytest.raises(
         ValueError, match=re.escape("The problem configuration has no starting point.")
     ):
-        benchmarking_problem.starting_points  # noqa: B018
+        problem_configuration.starting_points  # noqa: B018
 
 
 def check_default_starting_point(
-    benchmarking_problem: BaseBenchmarkingProblem, variable_space: DesignSpace
+    problem_configuration: BaseProblemConfiguration, variable_space: DesignSpace
 ) -> None:
     """Check the default starting point.
 
     Args:
-        benchmarking_problem: The benchmarking problem.
-        variable_space: The variable space of the benchmarking problem.
+        problem_configuration: The problem configuration.
+        variable_space: The variable space of the problem configuration.
     """
-    (starting_point,) = benchmarking_problem.starting_points
+    (starting_point,) = problem_configuration.starting_points
     assert (starting_point == variable_space.get_current_value()).all()
 
 
 def check_inconsistent_starting_points(
-    benchmarking_problem_class: type[BaseBenchmarkingProblem],
+    problem_configuration_class: type[BaseProblemConfiguration],
     create_problem: Callable,
     *args: Any,
 ) -> None:
     """Check initialization with starting points of inadequate size.
 
     Args:
-        benchmarking_problem_class: The class of benchmarking problem.
+        problem_configuration_class: The class of problem configuration.
         create_problem: A function that creates problems.
         *args: Positional arguments for the class initialization.
     """
@@ -89,27 +89,27 @@ def check_inconsistent_starting_points(
             "A starting point must be a 1-dimensional NumPy array of size 2."
         ),
     ):
-        benchmarking_problem_class(
+        problem_configuration_class(
             "Problem", create_problem, *args, starting_points=[numpy.zeros(3)]
         )
 
 
 def check_undefined_target_values(
-    benchmarking_problem: BaseBenchmarkingProblem,
+    problem_configuration: BaseProblemConfiguration,
 ) -> None:
     """Check that there is no target value.
 
     Args:
-        benchmarking_problem: The benchmarking problem.
+        problem_configuration: The problem configuration.
     """
     with pytest.raises(
         ValueError, match=re.escape("The problem configuration has no target value.")
     ):
-        benchmarking_problem.target_values  # noqa: B018
+        problem_configuration.target_values  # noqa: B018
 
 
 def __check_starting_points_generation(
-    benchmarking_problem_class: type[BaseBenchmarkingProblem],
+    problem_configuration_class: type[BaseProblemConfiguration],
     create_problem: Callable,
     input_doe_size: int | None,
     actual_doe_size: int,
@@ -118,7 +118,7 @@ def __check_starting_points_generation(
     """Check the generation of the starting points.
 
     Args:
-        benchmarking_problem_class: The class of benchmarking problem.
+        problem_configuration_class: The class of problem configuration.
         create_problem: A function that creates problems.
         input_doe_size: The input number of starting points.
         actual_doe_size: The actual number of starting points.
@@ -126,7 +126,7 @@ def __check_starting_points_generation(
     """
     assert (
         len(
-            benchmarking_problem_class(
+            problem_configuration_class(
                 "Problem",
                 create_problem,
                 *args,
@@ -139,36 +139,36 @@ def __check_starting_points_generation(
 
 
 def check_starting_points_default_generation(
-    benchmarking_problem_class: type[BaseBenchmarkingProblem],
+    problem_configuration_class: type[BaseProblemConfiguration],
     create_problem: Callable,
     *args: Any,
 ) -> None:
     """Check the generation of the starting points.
 
     Args:
-        benchmarking_problem_class: The class of benchmarking problem.
+        problem_configuration_class: The class of problem configuration.
         create_problem: A function that creates problems.
         *args: Positional arguments for the class initialization.
     """
     __check_starting_points_generation(
-        benchmarking_problem_class, create_problem, None, 2, *args
+        problem_configuration_class, create_problem, None, 2, *args
     )
 
 
 def check_starting_points_generation(
-    benchmarking_problem_class: type[BaseBenchmarkingProblem],
+    problem_configuration_class: type[BaseProblemConfiguration],
     create_problem: Callable,
     *args: Any,
 ) -> None:
     """Check the generation of the starting points.
 
     Args:
-        benchmarking_problem_class: The class of benchmarking problem.
+        problem_configuration_class: The class of problem configuration.
         create_problem: A function that creates problems.
         *args: Positional arguments for the class initialization.
     """
     __check_starting_points_generation(
-        benchmarking_problem_class, create_problem, 3, 3, *args
+        problem_configuration_class, create_problem, 3, 3, *args
     )
 
 
@@ -179,42 +179,42 @@ message = (
 
 
 def check_set_starting_points_as_non_2d_array(
-    benchmarking_problem: BaseBenchmarkingProblem,
+    problem_configuration: BaseProblemConfiguration,
 ) -> None:
     """Check the setting of starting points as a non 2-dimensional NumPy array.
 
     Args:
-        benchmarking_problem: The benchmarking problem.
+        problem_configuration: The problem configuration.
     """
     with pytest.raises(
         ValueError,
         match=re.escape(f"{message} A 1-dimensional NumPy array was passed."),
     ):
-        benchmarking_problem.starting_points = zeros(2)
+        problem_configuration.starting_points = zeros(2)
 
 
 def check_set_starting_points_as_non_iterable(
-    benchmarking_problem: BaseBenchmarkingProblem,
+    problem_configuration: BaseProblemConfiguration,
 ) -> None:
     """Check the setting of starting points as a non-iterable.
 
     Args:
-        benchmarking_problem: The benchmarking problem.
+        problem_configuration: The problem configuration.
     """
     with pytest.raises(
         TypeError,
         match=re.escape(f"{message} The following type was passed: {float}."),
     ):
-        benchmarking_problem.starting_points = 0.0
+        problem_configuration.starting_points = 0.0
 
 
 def check_set_starting_points_with_wrong_dimension(
-    benchmarking_problem: BaseBenchmarkingProblem,
+    problem_configuration: BaseProblemConfiguration,
 ) -> None:
     """Check the setting of starting points of the wrong dimension.
 
     Args:
-        benchmarking_problem: The benchmarking problem.
+        problem_configuration: The problem configuration.
     """
     with pytest.raises(
         ValueError,
@@ -223,7 +223,7 @@ def check_set_starting_points_with_wrong_dimension(
             "dimension (2)."
         ),
     ):
-        benchmarking_problem.starting_points = numpy.zeros((3, 1))
+        problem_configuration.starting_points = numpy.zeros((3, 1))
 
 
 test_description_parametrize = pytest.mark.parametrize(
@@ -240,7 +240,7 @@ test_description_parametrize = pytest.mark.parametrize(
 
 @test_description_parametrize
 def check_description(
-    benchmarking_problem_class: type[BaseBenchmarkingProblem],
+    problem_configuration_class: type[BaseProblemConfiguration],
     create_problem: Callable,
     input_description: Mapping[str, str],
     actual_description: str,
@@ -249,14 +249,14 @@ def check_description(
     """Check the description.
 
     Args:
-        benchmarking_problem_class: The class of benchmarking problem.
+        problem_configuration_class: The class of problem configuration.
         create_problem: A function that creates problems.
         input_description: The input description.
         actual_description: The actual description.
         *args: Positional arguments for the class initialization.
     """
     assert (
-        benchmarking_problem_class(
+        problem_configuration_class(
             "Problem", create_problem, *args, **input_description
         ).description
         == actual_description
@@ -265,7 +265,7 @@ def check_description(
 
 def check_starting_points_saving(
     tmp_path,
-    benchmarking_problem_class: type[BaseBenchmarkingProblem],
+    problem_configuration_class: type[BaseProblemConfiguration],
     create_problem: Callable,
     *args: Any,
 ) -> None:
@@ -273,13 +273,13 @@ def check_starting_points_saving(
 
     Args:
         tmp_path: The path to a temporary directory.
-        benchmarking_problem_class: The class of benchmarking problem.
+        problem_configuration_class: The class of problem configuration.
         create_problem: A function that creates problems.
         *args: Positional arguments for the class initialization.
     """
     starting_points = numpy.ones((3, 2))
     path = tmp_path / "starting_points.npy"
-    benchmarking_problem_class(
+    problem_configuration_class(
         "Problem", create_problem, *args, starting_points=starting_points
     ).save_starting_points(path)
     assert_equal(numpy.load(path), starting_points)
@@ -287,7 +287,7 @@ def check_starting_points_saving(
 
 def check_starting_point_loading(
     tmp_path,
-    benchmarking_problem_class: type[BaseBenchmarkingProblem],
+    problem_configuration_class: type[BaseProblemConfiguration],
     create_problem: Callable,
     *args: Any,
 ) -> None:
@@ -295,14 +295,14 @@ def check_starting_point_loading(
 
     Args:
         tmp_path: The path to a temporary directory.
-        benchmarking_problem_class: The class of benchmarking problem.
+        problem_configuration_class: The class of problem configuration.
         create_problem: A function that creates problems.
         *args: Positional arguments for the class initialization.
     """
     starting_points = numpy.ones((3, 2))
     path = tmp_path / "starting_points.npy"
     numpy.save(path, starting_points)
-    problem = benchmarking_problem_class(
+    problem = problem_configuration_class(
         "problem", create_problem, *args, starting_points=starting_points
     )
     problem.load_starting_point(path)
@@ -322,7 +322,7 @@ test_compute_data_profiles_parametrize = pytest.mark.parametrize(
 
 
 def check_data_profiles_computation(
-    benchmarking_problem_class: type[BaseBenchmarkingProblem],
+    problem_configuration_class: type[BaseProblemConfiguration],
     create_problem: Callable,
     target_values: TargetValues,
     algorithm_configurations: AlgorithmsConfigurations,
@@ -333,7 +333,7 @@ def check_data_profiles_computation(
     """Check the computation of data profiles.
 
     Args:
-        benchmarking_problem_class: The class of benchmarking problem.
+        problem_configuration_class: The class of problem configuration.
         create_problem: A function that creates problems.
         target_values: The target values.
         algorithm_configurations: The algorithm configurations.
@@ -346,7 +346,7 @@ def check_data_profiles_computation(
     target_values.compute_target_hits_history = mock.Mock(
         return_value=[0, 0, 0, 1, 1, 2]
     )
-    benchmarking_problem_class(
+    problem_configuration_class(
         "Problem", create_problem, *args, target_values=target_values
     ).compute_data_profile(
         algorithm_configurations,
@@ -356,7 +356,7 @@ def check_data_profiles_computation(
 
 
 def check_data_profiles_computation_with_max_eval_number(
-    benchmarking_problem_class: type[BaseBenchmarkingProblem],
+    problem_configuration_class: type[BaseProblemConfiguration],
     create_problem: Callable,
     target_values: TargetValues,
     algorithm_configurations: AlgorithmsConfigurations,
@@ -366,7 +366,7 @@ def check_data_profiles_computation_with_max_eval_number(
     """Check the computation of data profiles when the evaluations number is limited.
 
     Args:
-        benchmarking_problem_class: The class of benchmarking problem.
+        problem_configuration_class: The class of problem configuration.
         create_problem: A function that creates problems.
         target_values: The target values.
         algorithm_configurations: The algorithm configurations.
@@ -375,7 +375,7 @@ def check_data_profiles_computation_with_max_eval_number(
     """
     matplotlib.pyplot.close("all")
     target_values.compute_target_hits_history = mock.Mock(return_value=[0, 0, 0, 1])
-    bench_problem = benchmarking_problem_class(
+    bench_problem = problem_configuration_class(
         "Problem", create_problem, *args, target_values=target_values
     )
     bench_problem.compute_data_profile(
@@ -384,27 +384,27 @@ def check_data_profiles_computation_with_max_eval_number(
 
 
 def check_variable_space(
-    benchmarking_problem: BaseBenchmarkingProblem, variable_space: DesignSpace
+    problem_configuration: BaseProblemConfiguration, variable_space: DesignSpace
 ) -> None:
     """Check the variable space.
 
     Args:
-        benchmarking_problem: The benchmarking problem.
-        variable_space: The variable space of the benchmarking problem.
+        problem_configuration: The problem configuration.
+        variable_space: The variable space of the problem configuration.
     """
-    assert benchmarking_problem.variable_space is variable_space
+    assert problem_configuration.variable_space is variable_space
 
 
 def check_worker_type(
-    benchmarking_problem: BaseBenchmarkingProblem, worker_type: type[BaseWorker]
+    problem_configuration: BaseProblemConfiguration, worker_type: type[BaseWorker]
 ) -> None:
     """Check the type of benchmarking worker.
 
     Args:
-        benchmarking_problem_class: The class of benchmarking problem.
+        problem_configuration_class: The class of problem configuration.
         worker_type: The type of benchmarking worker.
     """
-    assert benchmarking_problem.worker is worker_type
+    assert problem_configuration.worker is worker_type
 
 
 @pytest.fixture